added prices for weapon upgrades

+++ Assets/_Game/Scripts/Components/Weapons/LightningShooterComponent.cs
@@ -5,7 +5,7 @@ public class LightningShooterComponent : IComponent
 {
     public float Cooldown;
     public float EffectRadius;
-    public int TargetDamage;
+    public float TargetDamage;
     public float AOEDamage;
     public float StunDuration;
 }
